# src/main/resources/static/openapi.yaml
openapi: 3.1.0
info:
  title: Bookstack backend API
  description: Library management API
  version: 1.0.0

servers:
  - url: 'http://localhost:8080'

tags:
  - name: Books
  - name: Authors
  - name: Genres
  - name: Publishers
  - name: BookCopies
  - name: Stock

components:
  schemas:
    Format:
      type: string
      enum: [ PAPERBACK, HARDBACK, POCKET ]
      description: Physical format of a copy

    Language:
      type: string
      enum: [ ENG, FR, GER, MG, OTHER ]
      description: Language of a book copy or an author

    Status:
      type: string
      enum: [ PENDING, COMPLETED, CANCELLED, BOOKED ]
      description: Sales status

    StockStatus:
      type: string
      enum: [ AVAILABLE, LOW, OUT_OF_STOCK ]
      description: Stock Status of a book copy

    Role:
      type: string
      enum: [ LIBRARIAN, ASSISTANT_LIBRARIAN, LIBRARY_MANAGER, ARCHIVIST, CLERK ]
      description: Role of an administrator

    PaymentMethod:
      type: string
      enum: [ CASH, CARD, MOBILE_MONEY ]
      description: Payment method

    Genre:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        name:
          type: string
          example: Romance
        description:
          type: string
          example: Stories centered around love and relationships.

    Publisher:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        name:
          type: string
          example: Gallimard
        email:
          type: string
          format: email
          example: contact@gallimard.fr
        website:
          type: string
          format: uri
          example: https://www.gallimard.fr
        location:
          type: string
          example: Paris, France

    Author:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        firstName:
          type: string
          example: Antoine
        lastName:
          type: string
          example: de Saint-Exupéry
        biography:
          type: string
          example: Écrivain et aviateur français.
        language:
          $ref: '#/components/schemas/Language'
        email:
          type: string
          format: email
          example: saint.exupery@example.com

    PriceHistory:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        price:
          type: number
          format: decimal
          example: 12.99
        changedAt:
          type: string
          format: date-time
          example: 2026-07-1T10:30:00Z

    Book:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        title:
          type: string
          example: Le Petit Prince
        summary:
          type: string
          example: Un petit prince voyage de planète en planète et apprend des leçons sur la vie.
        isbn:
          type: string
          example: 978-2070612758
        authors:
          type: array
          items:
            $ref: '#/components/schemas/Author'
        genres:
          type: array
          items:
            $ref: '#/components/schemas/Genre'

    BookCopy:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        format:
          $ref: '#/components/schemas/Format'
        language:
          $ref: '#/components/schemas/Language'
        book:
          $ref: '#/components/schemas/Book'
        publisher:
          $ref: '#/components/schemas/Publisher'
        priceHistories:
          type: array
          items:
            $ref: '#/components/schemas/PriceHistory'


    BookRequest:
      type: object
      required: [ title, summary, isbn ]
      properties:
        title:
          type: string
          example: Les Misérables
        summary:
          type: string
          example: Roman historique de Victor Hugo publié en 1862.
        isbn:
          type: string
          example: 978-2-07-040850-4
        authorIds:
          type: array
          items:
            type: string
            format: uuid
          example: [ "a1a1a1a1-..." ]
        genreIds:
          type: array
          items:
            type: string
            format: uuid
          example: [ "e4b6a7c8-..." ]

    BookCopyRequest:
      type: object
      required: [ format, language, bookId, publisherId ]
      properties:
        format:
          $ref: '#/components/schemas/Format'
        language:
          $ref: '#/components/schemas/Language'
        bookId:
          type: string
          format: uuid
        publisherId:
          type: string
          format: uuid


    FormatStock:
      type: object
      description: Stock of a specific BookCopy edition
      properties:
        bookCopyId:
          type: string
          format: uuid
        format:
          $ref: '#/components/schemas/Format'
        language:
          $ref: '#/components/schemas/Language'
        stock:
          type: integer
          example: 5

    BookStock:
      type: object
      description: Stock of a book all editions combined
      properties:
        bookId:
          type: string
          format: uuid
        title:
          type: string
          example: Les Misérables
        isbn:
          type: string
          example: 978-2-07-040850-4
        totalStock:
          type: integer
          description: Total inventory of all editions
          example: 13
        formats:
          type: array
          description: Stock details by edition
          items:
            $ref: '#/components/schemas/FormatStock'

    StockResponse:
      type: object
      description: Stock of a BookCopy with its status
      properties:
        bookCopyId:
          type: string
          format: uuid
        bookCopyTitle:
          type: string
          example: Les Misérables — Poche FR
        stock:
          type: integer
          example: 2
        status:
          $ref: '#/components/schemas/StockStatus'

    GenreRevenue:
      type: object
      description: Revenue generated by a genre over a given period
      properties:
        genreId:
          type: string
          format: uuid
        genreName:
          type: string
          example: Romance
        totalRevenue:
          type: number
          format: decimal
          description: Total COMPLETED sales for this genre during the period
          example: 4520.50
        periodStart:
          type: string
          format: date-time
          example: 2025-07-01T00:00:00Z
        periodEnd:
          type: string
          format: date-time
          example: 2025-07-07T23:59:59Z

    BookInfoDTO:
      type: object
      description: Book information from external sources (Google Books, Open Library)
      properties:
        title:
          type: string
          example: Le Petit Prince
        authors:
          type: string
          example: Antoine de Saint-Exupéry
        isbn:
          type: string
          example: 978-2070612758
        source:
          type: string
          description: Data source (Google Books or Open Library)
          example: GOOGLE
        description:
          type: string
          example: Un petit prince voyage de planète en planète...
        publishDate:
          type: string
          example: 1943-04-06
        pageCount:
          type: integer
          example: 96
        imageUrl:
          type: string
          format: uri
          example: "https://books.google.com/books/content?id=..."
        openLibraryUrl:
          type: string
          format: uri
          example: "https://openlibrary.org/books/OL..."


    ErrorBody:
      type: object
      properties:
        status:
          type: integer
          example: 404
        error:
          type: string
          example: Not Found
        message:
          type: string
          example: Book with id xxx not found
        path:
          type: string
          example: /books/xxx
        timestamp:
          type: string
          format: date-time


  responses:
    400BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'

    404NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'

    409Conflict:
      description: The resource already exists
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'

    500InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'


paths:
  /books:
    get:
      tags: [ Books ]
      summary: List all books
      responses:
        200:
          description: List of all books
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Book'
        500:
          $ref: '#/components/responses/500InternalError'

    post:
      tags: [ Books ]
      summary: Create one or more books
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/BookRequest'
      responses:
        201:
          description: Created books
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Book'
        400:
          $ref: '#/components/responses/400BadRequest'
        409:
          $ref: '#/components/responses/409Conflict'
        500:
          $ref: '#/components/responses/500InternalError'


  /books/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      tags: [ Books ]
      summary: Get book by its id
      responses:
        '200':
          description: Book found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
        404:
          $ref: '#/components/responses/404NotFound'
        500:
          $ref: '#/components/responses/500InternalError'

    put:
      tags: [ Books ]
      summary: Update a book
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookRequest'
      responses:
        200:
          description: Book updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
        400:
          $ref: '#/components/responses/400BadRequest'
        404:
          $ref: '#/components/responses/404NotFound'
        409:
          $ref: '#/components/responses/409Conflict'
        500:
          $ref: '#/components/responses/500InternalError'

    delete:
      tags: [ Books ]
      summary: Delete a book
      responses:
        204:
          description: Book deleted
        404:
          $ref: '#/components/responses/404NotFound'
        500:
          $ref: '#/components/responses/500InternalError'

  /books/search:
    get:
      tags: [ Books ]
      summary: Search book by ISBN from external sources
      description: Search for book information using ISBN from Google Books and Open Library APIs
      parameters:
        - name: isbn
          in: query
          required: true
          schema:
            type: string
            example: 978-2070612758
          description: ISBN number of the book
      responses:
        200:
          description: Book information found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookInfoDTO'
        400:
          $ref: '#/components/responses/400BadRequest'
        404:
          description: Book not found in external sources
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookInfoDTO'
                example:
                  source: NOT_FOUND
                  isbn: 978-2070612758
        500:
          $ref: '#/components/responses/500InternalError'

  
  /book-copy:
    get:
      tags: [ BookCopies ]
      summary: List of all BookCopies
      responses:
        '200':
          description: BookCopies list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BookCopy'
        '500':
          $ref: '#/components/responses/500InternalError'

    post:
      tags: [ BookCopies ]
      summary: Create one or more BookCopy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/BookCopyRequest'
      responses:
        '200':
          description: BookCopies created
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BookCopy'
        '400':
          $ref: '#/components/responses/400BadRequest'
        '404':
          $ref: '#/components/responses/404NotFound'
        '500':
          $ref: '#/components/responses/500InternalError'


  /book-copy/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      tags: [ BookCopies ]
      summary: Get BookCopy by its id
      responses:
        '200':
          description: BookCopy found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookCopy'
        '404':
          $ref: '#/components/responses/404NotFound'
        '500':
          $ref: '#/components/responses/500InternalError'


  /stock/book/{bookId}:
    parameters:
      - name: bookId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [ Stock ]
      summary: Stock of a book all editions combined
      description: Returns the total stock and a breakdown by edition for a given book
      responses:
        200:
          description: Book stock
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookStock'
        '404':
          $ref: '#/components/responses/404NotFound'
        '500':
          $ref: '#/components/responses/500InternalError'

  /stock/format/{bookCopyId}:
    parameters:
      - name: bookCopyId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [ Stock ]
      summary: Stock of a specific edition
      description: Returns the stock for a specific BookCopy
      responses:
        200:
          description: Edition stock
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FormatStock'
        404:
          $ref: '#/components/responses/404NotFound'
        500:
          $ref: '#/components/responses/500InternalError'

  /stock/low:
    get:
      tags: [ Stock ]
      summary: Books with low stock
      description: >
        Returns all books whose total stock is less than or equal to the
        threshold (default=3).
      parameters:
        - name: threshold
          in: query
          required: false
          schema:
            type: integer
            default: 3
            minimum: 0
          example: 3
      responses:
        200:
          description: List of books with low stock
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BookStock'
        400:
          $ref: '#/components/responses/400BadRequest'
        500:
          $ref: '#/components/responses/500InternalError'


  /authors:
    get:
      tags: [ Authors ]
      summary: List all authors
      responses:
        200:
          description: List of all authors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Author'
        500:
          $ref: '#/components/responses/500InternalError'

    post:
      tags: [ Authors ]
      summary: Create one or more authors
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/Author'
      responses:
        200:
          description: Authors created
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Author'
        400:
          $ref: '#/components/responses/400BadRequest'
        409:
          $ref: '#/components/responses/409Conflict'
        500:
          $ref: '#/components/responses/500InternalError'

  /authors/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      tags: [ Authors ]
      summary: Get author by his/her id
      operationId: getAuthorById
      responses:
        200:
          description: Auteur trouvé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Author'
        404:
          $ref: '#/components/responses/404NotFound'
        500:
          $ref: '#/components/responses/500InternalError'

    put:
      tags: [ Authors ]
      summary: Update an author
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Author'
      responses:
        200:
          description: Author updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Author'
        400:
          $ref: '#/components/responses/400BadRequest'
        404:
          $ref: '#/components/responses/404NotFound'
        500:
          $ref: '#/components/responses/500InternalError'

    delete:
      tags: [ Authors ]
      summary: Delete an author
      responses:
        204:
          description: Author deleted
        404:
          $ref: '#/components/responses/404NotFound'
        500:
          $ref: '#/components/responses/500InternalError'


  /genres:
    get:
      tags: [ Genres ]
      summary: List all genres
      responses:
        200:
          description: List of all genres
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Genre'
        500:
          $ref: '#/components/responses/500InternalError'

    post:
      tags: [ Genres ]
      summary: Create one or more genres
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/Genre'
      responses:
        200:
          description: Genres created
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Genre'
        400:
          $ref: '#/components/responses/400BadRequest'
        409:
          $ref: '#/components/responses/409Conflict'
        500:
          $ref: '#/components/responses/500InternalError'


  /genres/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      tags: [ Genres ]
      summary: Get genre by its id
      responses:
        200:
          description: Genre found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Genre'
        404:
          $ref: '#/components/responses/404NotFound'
        500:
          $ref: '#/components/responses/500InternalError'

    put:
      tags: [ Genres ]
      summary: Update a genre
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Genre'
      responses:
        200:
          description: Genre updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Genre'
        400:
          $ref: '#/components/responses/400BadRequest'
        404:
          $ref: '#/components/responses/404NotFound'
        500:
          $ref: '#/components/responses/500InternalError'

    delete:
      tags: [ Genres ]
      summary: Delete a genre
      responses:
        204:
          description: Genre deleted
        404:
          $ref: '#/components/responses/404NotFound'
        500:
          $ref: '#/components/responses/500InternalError'

  /genres/revenue:
    parameters:
      - name: genre
        in: query
        required: true
        schema:
          type: string
          format: string
    get:
      tags: [ Genres ]
      summary: Revenue generated by a genre over a period of time
      description: >
        Returns the total COMPLETED sales for books belonging
        to this genre, for the period defined by 'start' and 'end'.
      responses:
        200:
          description: Revenus du genre sur la période
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenreRevenue'
        400:
          $ref: '#/components/responses/400BadRequest'
        404:
          $ref: '#/components/responses/404NotFound'
        500:
          $ref: '#/components/responses/500InternalError'


  /publishers:
    get:
      tags: [ Publishers ]
      summary: List all publishers
      responses:
        200:
          description: List of all publishers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Publisher'
        500:
          $ref: '#/components/responses/500InternalError'

    post:
      tags: [ Publishers ]
      summary: Create one or more publisher
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/Publisher'
      responses:
        200:
          description: Publishers created
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Publisher'
        400:
          $ref: '#/components/responses/400BadRequest'
        409:
          $ref: '#/components/responses/409Conflict'
        500:
          $ref: '#/components/responses/500InternalError'

  /publishers/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      tags: [ Publishers ]
      summary: Get a publisher by his/her id
      responses:
        200:
          description: Publisher found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Publisher'
        404:
          $ref: '#/components/responses/404NotFound'
        500:
          $ref: '#/components/responses/500InternalError'

    put:
      tags: [ Publishers ]
      summary: Update a publisher
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Publisher'
      responses:
        200:
          description: Publisher updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Publisher'
        400:
          $ref: '#/components/responses/400BadRequest'
        404:
          $ref: '#/components/responses/404NotFound'
        500:
          $ref: '#/components/responses/500InternalError'

    delete:
      tags: [ Publishers ]
      summary: Delete a publisher
      responses:
        204:
          description: Publisher deleted
        404:
          $ref: '#/components/responses/404NotFound'
        500:
          $ref: '#/components/responses/500InternalError'